# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Netpro",
        "description": "Dokumentasi API untuk Netpro",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/mikrotik/set-profile": {
            "post": {
                "tags": [
                    "Hotspot Profile"
                ],
                "summary": "Create or link a hotspot profile",
                "description": "Membuat hotspot profile baru atau menambahkan link jika profile sudah ada di Mikrotik.",
                "operationId": "b93078f815e80aa6d88f35fdd9284f43",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Tenant-ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "profile_name",
                                    "shared_users"
                                ],
                                "properties": {
                                    "profile_name": {
                                        "type": "string",
                                        "example": "Premium-User"
                                    },
                                    "shared_users": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "rate_limit": {
                                        "type": "string",
                                        "example": "2M/2M"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/package"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotspot profile created successfully"
                    },
                    "200": {
                        "description": "Profile already exists but link added"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mikrotik/get-profile": {
            "get": {
                "tags": [
                    "Hotspot Profile"
                ],
                "summary": "Get all hotspot profiles",
                "description": "Mengambil semua data hotspot profile yang ada di Mikrotik dan link yang tersimpan di database.",
                "operationId": "9ffbe176f572d73b54371092efe9f5de",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Tenant-ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profiles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profiles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "profile_name": {
                                                        "type": "string",
                                                        "example": "Premium-User"
                                                    },
                                                    "shared_users": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "rate_limit": {
                                                        "type": "string",
                                                        "example": "2M/2M"
                                                    },
                                                    "link": {
                                                        "type": "string",
                                                        "example": "https://example.com/package"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No profiles found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mikrotik/get-profile/{profile_name}": {
            "get": {
                "tags": [
                    "Hotspot Profile"
                ],
                "summary": "Get hotspot profile by name",
                "description": "Mengambil data hotspot profile berdasarkan nama.",
                "operationId": "5219b6b9d904b59083e7b65c15089013",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Tenant-ID"
                    },
                    {
                        "name": "profile_name",
                        "in": "path",
                        "description": "Nama profile yang ingin dicari",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile data"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mikrotik/delete-profile/{profile_name}": {
            "delete": {
                "tags": [
                    "Hotspot Profile"
                ],
                "summary": "Delete hotspot profile",
                "description": "Menghapus hotspot profile berdasarkan nama di Mikrotik.",
                "operationId": "c6efe4b350c38dbf8f38bc35589f00c9",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Tenant-ID"
                    },
                    {
                        "name": "profile_name",
                        "in": "path",
                        "description": "Nama profile yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotspot profile deleted successfully"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mikrotik/hotspot-profile/{profile_name}": {
            "put": {
                "tags": [
                    "Hotspot Profile"
                ],
                "summary": "Update hotspot profile",
                "description": "Mengupdate data hotspot profile dan link di database.",
                "operationId": "d92578f75cf1e8fd8100026dd85be2f1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Tenant-ID"
                    },
                    {
                        "name": "profile_name",
                        "in": "path",
                        "description": "Nama profile yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "link"
                                ],
                                "properties": {
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/package"
                                    },
                                    "shared_users": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "rate_limit": {
                                        "type": "string",
                                        "example": "3M/3M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotspot profile and link updated successfully"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "X-Tenant-ID": {
                "name": "X-Tenant-ID",
                "in": "header",
                "description": "Encrypted Tenant ID dari hasil login",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "eyJpdiI6IjE2MjM..."
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan Bearer Token untuk autentikasi.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}